requirements.txt
Pillow>=8.0.1
numpy>=1.19.4
setuptools_rust>=0.11.6
pytest